requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
tensorboard>=2.4.0
scipy>=1.5.0
tqdm>=4.50.0
matplotlib>=3.3.0
pyyaml>=5.3.0

